+++ src/pages/LoadUnload/components/MapLayers/MapLayers.tsx
@@ -1,21 +1,33 @@
 import { BaseLayer } from '@amsterdam/arm-core'
 
 import { AddressMarker } from '../../../../shared/components/MapLayers/AddressMarker'
+import { HighlightedFeatureLayer } from '../../../../shared/components/HighlightedFeatureLayer'
 import { topoBlackWhite } from '../../../../shared/map/mapLayers'
 
+import { useLoadUnloadMapContext } from '../../contexts/MapContext'
 import { useLoadUnloadPageContext } from '../../contexts/PageContext'
 
-import { LoadUnloadHighlightedFeatureLayer } from './HighlightedFeatureLayer'
+import { LoadUnloadRoadSectionsLoadUnloadLayer } from './RoadSectionsLoadUnloadLayer'
 import { LoadUnloadLoadUnloadSpacesLayer } from './LoadUnloadSpacesLayer'
 
 export const LoadUnloadMapLayers = () => {
+  const { detailFeature } = useLoadUnloadMapContext()
   const { address } = useLoadUnloadPageContext()
 
   return (
     <>
       <AddressMarker address={address} />
 
-      <LoadUnloadHighlightedFeatureLayer />
+      {detailFeature.feature && (
+        <HighlightedFeatureLayer
+          detailFeature={{
+            id: detailFeature.feature.id,
+            geometry: detailFeature.feature.data.geometry,
+          }}
+        />
+      )}
+
+      <LoadUnloadRoadSectionsLoadUnloadLayer />
 
       <LoadUnloadLoadUnloadSpacesLayer />